complex::burning_ship_impl
Trait BurningShipOperations

pub trait BurningShipOperations {
    // Required methods
    fn new(c: Complex, divergence_threshold_square: f64) -> Self;
    fn divergence_threshold_square(&self) -> f64;
    fn c(&self) -> &Complex;
}

Provides operations specific to the BurningShip fractal.

Required Methods

fn new(c: Complex, divergence_threshold_square: f64) -> Self

Constructs a new BurningShipDescriptor with the specified complex number and divergence threshold.

fn divergence_threshold_square(&self) -> f64

Returns the square of the divergence threshold.

fn c(&self) -> &Complex

Returns a reference to the complex number c used in the Julia fractal formula.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.