complex::iterated_sinz_impl
Trait IteratedSinZOperations

pub trait IteratedSinZOperations {
    // Required methods
    fn new(c: Complex) -> Self;
    fn max_iteration(&self) -> u16;
    fn c(&self) -> &Complex;
    fn η(&self) -> f64;
}

Provides operations specific to the Iterated Sin(z) fractal.

Required Methods

fn new(c: Complex) -> Self

Constructs a new IteratedSinZDescriptor with the specified complex number and divergence threshold.

fn max_iteration(&self) -> u16

Fixed to 50

fn c(&self) -> &Complex

Returns a reference to the complex number c used in the IteratedSinZ fractal formula.

fn η(&self) -> f64

Returns the divergence threshold η used in the IteratedSinZ fractal formula.

Description

In the context of the Iterated SinZ fractal, this divergence threshold (η) is used to determine when the iterations should stop.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.